+++ game_core/__init__.py
@@ -1,0 +1,5 @@
+"""Initialization for the game_core package."""
+
+from .font_loader import FontManager, font_loader
+
+__all__ = ["FontManager", "font_loader"]
