+++ game_core/gameplay/other_components/scrollable_text_area.py
